refactor: move attempt functions into autoAction.js
